Http(s)URLConnection changes
Android 5.1 deprecates the Apache HttpClient classes, (see:
http://developer.android.com/about/versions/android-5.1.html#http) and
they are completely removed in the M developer preview. This means that
we should switch to using a different method for uploading NDK crashes
to HockeyApp.

This is not working yet, but I'm committing the new version
commented-out so it can be reviewed/improved upon.